delete 3 files and update 7 files
