+++ apps/mobile/src/navigation/ExploreStack.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 
 import { ExploreScreen } from '~/screens'
 
-import Header from '../components/common/Header'
+import Header from '../components/common/header/Header'
 
 const { Navigator, Screen } = createStackNavigator<ExploreStackParamList>()
