+++ apps/mobile/src/navigation/RootNavigator.tsx
@@ -6,7 +6,7 @@ import type { FC } from 'react'
 import React from 'react'
 import { Easing } from 'react-native'
 
-import CollapseButton from '~/components/common/CollapseButton'
+import CollapseButton from '~/components/common/header/CollapseButton'
 import {
   CategoriesModal,
   FeedFlexModal,
